some more moves
