--- src/editor/editor-state-prototypes.js
@@ -1,9 +1,0 @@
-import { stateSchema } from "./state-schemas/state-schema";
-
-import { newNotebookFromSchema } from "./state-schemas/state-prototype-from-schema";
-
-function newNotebook() {
-  return newNotebookFromSchema(stateSchema);
-}
-
-export { newNotebook, stateSchema };
