+++ src/editor/state-schemas/editor-state-prototypes.js
@@ -1,0 +1,9 @@
+import { stateSchema } from "./state-schema";
+
+import { newNotebookFromSchema } from "./state-prototype-from-schema";
+
+function newNotebook() {
+  return newNotebookFromSchema(stateSchema);
+}
+
+export { newNotebook, stateSchema };
