+++ src/eval-frame/store.js
@@ -4,7 +4,7 @@ import thunk from "redux-thunk";
 import { createLogger } from "redux-logger";
 
 import reducer from "./reducers/reducer";
-import { newNotebook } from "../editor/editor-state-prototypes";
+import { newNotebook } from "../editor/state-schemas/editor-state-prototypes";
 import evalFrameStateSelector from "../editor/state-schemas/eval-frame-state-selector";
 
 let enhancer;
